+++ Calculator/Substraction.py
@@ -1,2 +1,2 @@
 def subtraction(a, b):
-    return int(b) - int(a)
+    return float(b) - float(a)
